IloveJava

+++ src/main/java/StatClass/CountCalculator.java
@@ -4,7 +4,6 @@
  */
 package StatClass;
 
-import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 
@@ -12,23 +11,12 @@ import java.util.Map;
  *
  * @author user
  */
-//public class CountCalculator {
-//
-//    public static int[] calculateSampleSize(ArrayList<String[]> records) {
-//        int[] sampleSizes = new int[3];
-//        for (int i = 0; i < 3; i++) {
-//            sampleSizes[i] = records.size();
-//        }
-//        return sampleSizes;
-//    }
-//
-//}
 public class CountCalculator {
 
-    public static int[] calculateSampleSize(Map<String, List<Double>> dataMap) {
+    public static double[] calculateSampleSize(Map<String, List<Double>> dataMap) {
         int sampleSize = dataMap.size();
 
-        int[] sampleSizes = new int[sampleSize];
+        double[] sampleSizes = new double[sampleSize];
         int index = 0;
         for (List<Double> data : dataMap.values()) {
             sampleSizes[index] = data.size();
